+++ tests/test_rabbitmq.py
@@ -223,18 +223,6 @@ def test_rabbitmq_workers_handle_rabbit_failures_gracefully(rabbitmq_broker, rab
     assert sum(attempts) >= 1
 
 
-def test_rabbitmq_raises_an_exception_when_delaying_messages_for_too_long(rabbitmq_broker):
-    # Given that I have an actor
-    @dramatiq.actor
-    def do_nothing():
-        pass
-
-    # If I try to send it a delayed message farther than 7 days into the future
-    # I expect it to raise a value error
-    with pytest.raises(ValueError):
-        do_nothing.send_with_options(delay=7 * 86400 * 1000 + 1)
-
-
 def test_rabbitmq_connections_can_be_deleted_multiple_times(rabbitmq_broker):
     del rabbitmq_broker.connection
     del rabbitmq_broker.connection